requirements.txt
python-pptx
matplotlib
numpy
Pillow
requests
